Refactor word embedding feature code
